move Storage trait and implementations into its own module
